Move utils into correct parent directory
