+++ src/pages/commits/ActiveCommits/CommitChartTooltip.tsx
@@ -6,11 +6,10 @@ import { Disclaimer } from "@leafygreen-ui/typography";
 import { inactiveElementStyle } from "components/styles";
 import { taskStatusToCopy, mapTaskToBarchartColor } from "constants/task";
 import { size, zIndex } from "constants/tokens";
-
 import { TaskStatus } from "types/task";
-
 import { msToDuration } from "utils/string";
-import { ColorCount, getStatusesWithZeroCount } from "./utils";
+import { ColorCount } from "../utils";
+import { getStatusesWithZeroCount } from "./utils";
 
 const { gray } = palette;
 interface Props {
