+++ src/pages/commits/utils.test.ts
@@ -1,12 +1,18 @@
+import { palette } from "@leafygreen-ui/palette";
+import { taskStatusToCopy } from "constants/task";
 import { ProjectHealthView } from "gql/generated/types";
 import { TaskStatus } from "types/task";
+import { versions } from "./testData";
 import {
   getMainlineCommitsQueryVariables,
   getFilterStatus,
   impossibleMatch,
   FAILED_STATUSES,
   ALL_NON_FAILING_STATUSES,
+  getAllTaskStatsGroupedByColor,
 } from "./utils";
+
+const { red, green, yellow, gray, purple } = palette;
 
 describe("getFilterStatus", () => {
   it("should return an object containing booleans that describe what filters have been applied", () => {
@@ -717,6 +723,200 @@ describe("getMainlineCommitsQueryVariables", () => {
         variants: [],
         statuses: [],
       });
+    });
+  });
+});
+
+describe("getAllTaskStatsGroupedByColor", () => {
+  it("grab the taskStatusStats.statusCounts field from all versions, returns mapping between version id to its {grouped task stats, max, total}", () => {
+    expect(getAllTaskStatsGroupedByColor(versions)).toStrictEqual({
+      "12": {
+        stats: [
+          {
+            count: 8,
+            statuses: [
+              taskStatusToCopy[TaskStatus.TestTimedOut],
+              taskStatusToCopy[TaskStatus.Failed],
+            ],
+            color: red.base,
+            umbrellaStatus: TaskStatus.FailedUmbrella,
+            statusCounts: {
+              [TaskStatus.TestTimedOut]: 6,
+              [TaskStatus.Failed]: 2,
+            },
+          },
+          {
+            count: 7,
+            statuses: [
+              taskStatusToCopy[TaskStatus.SystemTimedOut],
+              taskStatusToCopy[TaskStatus.SystemUnresponsive],
+            ],
+            color: purple.dark2,
+            umbrellaStatus: TaskStatus.SystemFailureUmbrella,
+            statusCounts: {
+              [TaskStatus.SystemTimedOut]: 5,
+              [TaskStatus.SystemUnresponsive]: 2,
+            },
+          },
+          {
+            count: 4,
+            statuses: [taskStatusToCopy[TaskStatus.Dispatched]],
+            color: yellow.base,
+            umbrellaStatus: TaskStatus.RunningUmbrella,
+            statusCounts: { [TaskStatus.Dispatched]: 4 },
+          },
+          {
+            count: 2,
+            statuses: [taskStatusToCopy[TaskStatus.WillRun]],
+            color: gray.base,
+            umbrellaStatus: TaskStatus.ScheduledUmbrella,
+            statusCounts: { [TaskStatus.WillRun]: 2 },
+          },
+        ],
+        max: 8,
+        total: 21,
+      },
+      "13": {
+        stats: [
+          {
+            count: 6,
+            statuses: [taskStatusToCopy[TaskStatus.Succeeded]],
+            color: green.dark1,
+            umbrellaStatus: TaskStatus.Succeeded,
+            statusCounts: { [TaskStatus.Succeeded]: 6 },
+          },
+          {
+            count: 2,
+            statuses: [taskStatusToCopy[TaskStatus.Failed]],
+            color: red.base,
+            umbrellaStatus: TaskStatus.FailedUmbrella,
+            statusCounts: { [TaskStatus.Failed]: 2 },
+          },
+          {
+            count: 9,
+            statuses: [
+              taskStatusToCopy[TaskStatus.Dispatched],
+              taskStatusToCopy[TaskStatus.Started],
+            ],
+            color: yellow.base,
+            umbrellaStatus: TaskStatus.RunningUmbrella,
+            statusCounts: {
+              [TaskStatus.Dispatched]: 4,
+              [TaskStatus.Started]: 5,
+            },
+          },
+        ],
+        max: 9,
+        total: 17,
+      },
+      "14": {
+        stats: [
+          {
+            count: 4,
+            statuses: [taskStatusToCopy[TaskStatus.Succeeded]],
+            color: green.dark1,
+            umbrellaStatus: TaskStatus.Succeeded,
+            statusCounts: { [TaskStatus.Succeeded]: 4 },
+          },
+          {
+            count: 6,
+            statuses: [taskStatusToCopy[TaskStatus.TaskTimedOut]],
+            color: red.base,
+            umbrellaStatus: TaskStatus.FailedUmbrella,
+            statusCounts: { [TaskStatus.TaskTimedOut]: 6 },
+          },
+          {
+            count: 7,
+            statuses: [
+              taskStatusToCopy[TaskStatus.SystemFailed],
+              taskStatusToCopy[TaskStatus.SystemUnresponsive],
+            ],
+            color: purple.dark2,
+            umbrellaStatus: TaskStatus.SystemFailureUmbrella,
+            statusCounts: {
+              [TaskStatus.SystemFailed]: 5,
+              [TaskStatus.SystemUnresponsive]: 2,
+            },
+          },
+          {
+            count: 3,
+            statuses: [taskStatusToCopy[TaskStatus.SetupFailed]],
+            color: purple.light2,
+            umbrellaStatus: TaskStatus.SetupFailed,
+            statusCounts: { [TaskStatus.SetupFailed]: 3 },
+          },
+          {
+            count: 3,
+            statuses: [taskStatusToCopy[TaskStatus.Started]],
+            color: yellow.base,
+            umbrellaStatus: TaskStatus.RunningUmbrella,
+            statusCounts: { started: 3 },
+          },
+          {
+            count: 2,
+            statuses: [taskStatusToCopy[TaskStatus.Unscheduled]],
+            color: gray.dark1,
+            umbrellaStatus: TaskStatus.UndispatchedUmbrella,
+            statusCounts: { [TaskStatus.Unscheduled]: 2 },
+          },
+        ],
+        max: 7,
+        total: 25,
+      },
+      "123": {
+        stats: [
+          {
+            count: 4,
+            statuses: [taskStatusToCopy[TaskStatus.Succeeded]],
+            color: green.dark1,
+            umbrellaStatus: TaskStatus.Succeeded,
+            statusCounts: { [TaskStatus.Succeeded]: 4 },
+          },
+          {
+            count: 6,
+            statuses: [taskStatusToCopy[TaskStatus.TaskTimedOut]],
+            color: red.base,
+            umbrellaStatus: TaskStatus.FailedUmbrella,
+            statusCounts: { [TaskStatus.TaskTimedOut]: 6 },
+          },
+          {
+            count: 7,
+            statuses: [
+              taskStatusToCopy[TaskStatus.SystemFailed],
+              taskStatusToCopy[TaskStatus.SystemUnresponsive],
+            ],
+            color: purple.dark2,
+            umbrellaStatus: TaskStatus.SystemFailureUmbrella,
+            statusCounts: {
+              [TaskStatus.SystemFailed]: 5,
+              [TaskStatus.SystemUnresponsive]: 2,
+            },
+          },
+          {
+            count: 3,
+            statuses: [taskStatusToCopy[TaskStatus.SetupFailed]],
+            color: purple.light2,
+            umbrellaStatus: TaskStatus.SetupFailed,
+            statusCounts: { [TaskStatus.SetupFailed]: 3 },
+          },
+          {
+            count: 3,
+            statuses: [taskStatusToCopy[TaskStatus.Started]],
+            color: yellow.base,
+            umbrellaStatus: TaskStatus.RunningUmbrella,
+            statusCounts: { started: 3 },
+          },
+          {
+            count: 2,
+            statuses: [taskStatusToCopy[TaskStatus.Unscheduled]],
+            color: gray.dark1,
+            umbrellaStatus: TaskStatus.UndispatchedUmbrella,
+            statusCounts: { [TaskStatus.Unscheduled]: 2 },
+          },
+        ],
+        max: 7,
+        total: 25,
+      },
     });
   });
 });
